Move combinedparameter to its own module
